Change groupid org.hawkular.client to org.hawkular.metrics
This it can be used directly integrated in hawkular, instead of as a client
