+++ hawkular-dropwizard-reporter-factory/src/main/java/org/hawkular/metrics/dropwizard/HawkularReporterFactory.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.client.dropwizard;
+package org.hawkular.metrics.dropwizard;
 
 import java.util.Map;
