+++ hawkular-reporter-common/src/main/java/org/hawkular/metrics/reporter/http/HawkularHttpResponse.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.client.http;
+package org.hawkular.metrics.reporter.http;
 
 /**
  * Minimalist and API-agnostic response wrapper
